moved TestAPISpec back to test_metadata_api_live
